# docs/openapi-public.yaml
...
openapi: 3.0.3
info:
  title: ChangeNOW Dashboard API - public endpoints
  version: 7.2.3
  description: Generated from NestJS controllers. Includes only the latest version
    of each endpoint that is not protected by ApiKey*Guard. Endpoints with AuthGuard/JWT
    guards are marked with bearerAuth/cookieAuth security.
servers:
- url: "/"
  description: Relative base URL. Nest URI versioning is represented as /v{version}/...
    in paths.
...
paths:
...
  "/v1.1/transactions":
    post:
      tags:
      - tx-creation
      operationId: tx_creation_post_v1_1_transactions_createV1_1
      summary: createV1_1
      description: Public API endpoint not protected by ApiKey*Guard. Generated from
        NestJS controller metadata.
      x-source:
        file: apps/tx-creation/src/tx-creation.controller.ts
        line: 43
        guards:
        - JwtAuthGuard
        - Jwt2faAuthGuard
        - JwtCookieAuthGuard
        returnType: Observable<AxiosResponse>
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
      security:
      - bearerAuth: []
      - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CreateTxDto"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: cn_at
      x-legacy-cookie-name: accessToken
    csrfHeader:
      type: apiKey
      in: header
      name: x-csrf-token
      description: Required only for operations with CsrfGuard. The value must match
        the cn_csrf cookie.
  schemas:
...
    CreateTxDto:
      type: object
      properties:
        fromCurrency:
          type: string
        fromNetwork:
          type: string
        toCurrency:
          type: string
        toNetwork:
          type: string
        fromAmount:
          type: string
        toAmount:
          type: string
        address:
          type: string
        refundAddress:
          type: string
        extraId:
          type: string
        refundExtraId:
          type: string
        userId:
          type: string
        payload:
          type: string
        contactEmail:
          type: string
        flow:
          "$ref": "#/components/schemas/FlowTypes"
          default: FlowEnum.STANDARD
        type:
          "$ref": "#/components/schemas/EstimateTypeType"
          default: EstimateTypeEnum.DIRECT
        rateId:
          type: string
        value:
          type: string
        provider:
          "$ref": "#/components/schemas/TxProvidersType"
          default: ProvidersEnum.DEFAULT
        linkId:
          type: string
        message:
          type: string
        promoCode:
          type: string
        info:
          "$ref": "#/components/schemas/InfoDto"
        isDeposit:
          type: boolean
          default: 'false'
        brokerApiKey:
          type: string
        isBridge:
          type: boolean
          default: 'false'
        isTopUp:
          type: boolean
          default: 'false'
      required:
      - fromCurrency
      - fromNetwork
      - toCurrency
      - toNetwork
      - address
      - value
      x-source: apps/main/src/v2/dto/create-tx.dto.ts
...
    InfoDto:
      type: object
      properties:
        locale:
          type: string
        version:
          type: string
        platform:
          type: string
        utmSource:
          type: string
        utmMedium:
          type: string
        utmCampaign:
          type: string
        utmContent:
          type: string
        utmTerm:
          type: string
        landingPage:
          type: string
        exchangePage:
          type: string
        sessionId:
          type: string
        sessionCount:
          type: string
      x-source: apps/main/src/analytics/dto/info.dto.ts
...